% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_dt}
\alias{aggregate_dt}
\title{Aggregate values in a data.table object using a mapping.
If no weight is given, the value for the aggregated categories is the sum of the parts.
Otherwise, the weight is used to calculate a weighted average accross the parts.}
\usage{
aggregate_dt(data, mapping, fewcol = "region", yearcol = "year",
  manycol = "iso", datacols = "data", valuecol = "value",
  weights = NULL, weightcol = "weight")
}
\arguments{
\item{data, }{a magpie object.}

\item{mapping, }{a mapping between the aggregated categories in the data and ISO3 countrycodes. *All* regions in `data` have to be part of the mapping.}

\item{fewcol, }{name of the column containing aggregated categories. Default is "region".}

\item{yearcol, }{name of the column containing time step info. Default is "year".}

\item{manycol, }{name of the column containing dis-aggregated categories. Default is "iso".}

\item{datacols}{index columns that label categories which have to be treated seperately when aggregating with a weight.}

\item{valuecol}{name of the column with the value to aggregate, default is `value`.}

\item{weights}{table with weights for a (weighted average) aggregation, the name of the column with the aggregated categories has to be `manycol`. If columns (other than the column with the aggregated category) of the `weights` coincide with columns of the data, the respective columns are considered when joining.}

\item{weightcol}{column with the weights for aggregation, default is `weight`.}
}
\description{
Aggregate values in a data.table object using a mapping.
If no weight is given, the value for the aggregated categories is the sum of the parts.
Otherwise, the weight is used to calculate a weighted average accross the parts.
}
