% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getGDP_dt}
\alias{getGDP_dt}
\title{Load GDP data using `moinput` on ISO country resolution for a scenario as data.table object with given colnames.
By default, the}
\usage{
getGDP_dt(scenario = "gdp_SSP2", yearcol = "year", isocol = "iso",
  valuecol = "weight", usecache = F, gdpfile = "GDPcache.rds")
}
\arguments{
\item{scenario, }{GDP scenario, default is gdp_SSP2.}

\item{yearcol, }{name of the year column, default "year".}

\item{isocol, }{name of the column containing ISO3 codes, default is "iso".}

\item{valuecol, }{name of the column containing the GDP values, default is "weight".}

\item{usecache, }{store the result in a RDS file in the working directory, default is FALSE.}

\item{gdpfile, }{if caching is required, specify the filename here, default is "GDPcache.rds"}
}
\description{
Load GDP data using `moinput` on ISO country resolution for a scenario as data.table object with given colnames.
By default, the
}
\examples{
gdp <- getGDP_dt()
}
\keyword{gdp}
