% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{readMIF}
\alias{readMIF}
\title{Read a REMIND output (MIF) file.}
\usage{
readMIF(mif)
}
\arguments{
\item{mif}{A REMIND output file (.MIF)}
}
\description{
Note that these files are semi-colon separated CSVs with a trailing semi-colon
at the end of each entry.
}
\examples{
\dontrun{
dt <- readMIF("REMIND_generic_default.mif")
}
}
