% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REMIND_FinalEnergy}
\alias{REMIND_FinalEnergy}
\title{A random REMIND FE trajectory.}
\format{A data frame with 2011 rows and 6 columns:
\describe{
  \item{year}{REMIND time step}
  \item{region}{REMIND-12 region}
  \item{se}{Secondary energy identifier}
  \item{fe}{Final energy identifier}
  \item{te}{Conversion technology identifier}
  \item{value}{The data column, EJ/yr}
}}
\usage{
REMIND_FinalEnergy
}
\description{
A random REMIND FE trajectory.
}
\keyword{datasets}
