% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REMIND_GDP}
\alias{REMIND_GDP}
\title{REMIND GDP Trajectories for ISO countries}
\format{A data frame with 9462 rows and 4 columns:
\describe{
  \item{iso}{ISO3 country code}
  \item{variable}{SSP GDP variant, this is gdp_SSP2 in this case}
  \item{weight}{GDP in US$2010}
  \item{year}{Year}
}}
\usage{
REMIND_GDP
}
\description{
REMIND GDP Trajectories for ISO countries
}
\keyword{datasets}
