% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REMIND_RegionMap}
\alias{REMIND_RegionMap}
\title{The mapping between REMIND-12 regions and ISO countries}
\format{A data frame with 249 rows and 3 columns:
\describe{
  \item{name}{Name of the country}
  \item{iso}{ISO3 country code}
  \item{region}{REMIND region name}
}}
\usage{
REMIND_RegionMap
}
\description{
The mapping between REMIND-12 regions and ISO countries
}
\keyword{datasets}
