% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{toRegions_dt}
\alias{toRegions_dt}
\title{Aggregate values defined countrywise in a data.table object to regions using a mapping.
If the strategy is the default (`sum`), the value for the region is the sum of the (ISO3) country values.
If the strategy is `gdp`, GDP projections for the countries are used to calculate the regional value as a weighted average of the country values.}
\usage{
toRegions_dt(data, mapping, regioncol = "region", yearcol = "year",
  isocol = "iso", datacols = "data", valuecol = "value",
  strategy = "sum", scenario = "gdp_SSP2", usecache = F,
  gdpfile = "GDPcache.rds")
}
\arguments{
\item{data, }{a magpie object.}

\item{mapping, }{a mapping between the regions in the data and ISO3 countrycodes. *All* regions in `data` have to be part of the mapping.}

\item{regioncol, }{name of the column containing regions. A region is a set of countries. Default is "region".}

\item{yearcol, }{name of the column containing the year, default is "year".}

\item{datacols, }{index dimensions that are not regional or temporal dimensions.}

\item{valuecol, }{name of the column with the actual value to disaggregate, default is `value`.}

\item{strategy, }{if "gdp" is chosen, GDP projections for the countries are loaded and used as a weight to distribute the regional value.}

\item{scenario, }{the scenario for the GDP projections, default is `gdp_SSP2`.}

\item{usecache, }{if "gdp" is chosen as strategy, store the weight in a cachefile to speed up future executions.}

\item{gdpfile, }{if "gdp" is chosen as strategy and if caching is required, specify the filename here, default is "GDPcache.rds"}
}
\description{
Aggregate values defined countrywise in a data.table object to regions using a mapping.
If the strategy is the default (`sum`), the value for the region is the sum of the (ISO3) country values.
If the strategy is `gdp`, GDP projections for the countries are used to calculate the regional value as a weighted average of the country values.
}
