% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{varcalc_dt}
\alias{varcalc_dt}
\title{Execute *vertical* calculations along a given column.}
\usage{
varcalc_dt(dt, varcol, valcol, exp_str, res_var, ...)
}
\arguments{
\item{dt}{data.table, long format}

\item{varcol}{name of the column with the variable}

\item{valcol}{name of the column with the value}

\item{exp_str}{the expression to be calculated, a string}

\item{res_var}{name of the variable to which the result is to be stored}

\item{...}{other arguments are passed on to the data.table call where `exp` is evaluated.
Most likely you want to pass the `by=` parameter for group-by calls, see examples.}
}
\description{
This assumes a *long* format with a single value column,
dcasts the data.table to wide format, executes the calulation,
melts back to long format and returns the resulting data.table
with the additional column.
}
\details{
Note that the data.table should have at least three columns, i.e.,
the variable, the value and one id column.
}
\examples{
mt_dt <- as.data.table(mtcars, keep.rownames = TRUE)
## to long
mt1 <- melt(mt_dt, id.vars=c("rn", "cyl"))

varcalc_dt(mt1, "variable", "value", "wt/hp", "spec. hp")
varcalc_dt(mt1, "variable", "value", "sum(wt)", "wsum", by="cyl")
}
