% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{writeMIF}
\alias{writeMIF}
\title{Write a REMIND output (MIF) file.}
\usage{
writeMIF(dt, destination, append = FALSE, ...)
}
\arguments{
\item{dt}{a data.table in the correct format.}

\item{destination}{path to the resulting MIF file}

\item{append}{append to an existing MIF file?}

\item{...}{other parameters are passed on to data.table::fwrite}
}
\description{
Note that these files are semi-colon separated CSVs with a trailing semi-colon
at the end of each entry. Required columns are "Model", "Scenario",
"Region", "Variable", "Unit" and an arbitrary number of year colums
(should be convertable to numeric).
}
\details{
NAs are written to the file as "N/A".
}
\examples{
\dontrun{
writeMIF(dt, "REMIND_generic_default.mif")
}
}
