% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{readMIF}
\alias{readMIF}
\title{Read a REMIND output (MIF) file.}
\usage{
readMIF(mif)
}
\arguments{
\item{mif}{A REMIND output file (.MIF)}
}
\description{
REMIND style output files are semi-colon separated CSVs with a trailing semi-colon
at the end of each row. The following structure is assumed:
Columns "Model", "Scenario", "Region", "Variable", "Unit"
and an arbitrary number of year colums (convertable to numeric).
}
\examples{
\dontrun{
dt <- readMIF("REMIND_generic_default.mif")
}
}
