% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmndt-package.R
\docType{package}
\name{rmndt-package}
\alias{rmndt}
\alias{rmndt-package}
\title{rmndt: Tools for data.table objects in the REMIND context}
\description{
Helper functions for REMIND-related tasks with data.table objects, e.g., interpolation and (dis-)aggregation.
}
\author{
\strong{Maintainer}: Alois Dirnaichner \email{dirnaichner@pik-potsdam.de}

}
