% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDfToGDXparameter.R
\name{writeDfToGDXparameter}
\alias{writeDfToGDXparameter}
\title{Export Data Frame / Data Table to GDX Parameter and save to file
based on gamstransfer package}
\usage{
writeDfToGDXparameter(
  ddata,
  gdxFileName,
  paramName,
  domainCols = NULL,
  valueCol = "value",
  description = ""
)
}
\arguments{
\item{ddata}{data.frame or data.table}

\item{gdxFileName}{Output file name incl. path (e.g., "pXX_someName.gdx")}

\item{paramName}{Name of the GAMS parameter in GDX}

\item{domainCols}{Optional vector of column names for domains.
If NULL, defaults to all columns except `valueCol`.}

\item{valueCol}{Name of the numerical value column (default: "value")}

\item{description}{Optional description for the parameter (default: "")}
}
\value{
`NULL`, invisibly. Called for the side effect of writing `gdxFileName`.
}
\description{
Export Data Frame / Data Table to GDX Parameter and save to file
based on gamstransfer package
}
\author{
Alex K. Hagen
}
